Added FollowPath function
The followpath function using the ramsete controller was added. NOTE: Currently, an object is used to calculate and overlay the motion profile on the original path variable. Despite not being standard practice, this is done to 1. organize all the constraints in one place for better readibiltiy 2. eliminated the need to copy over an entire path and speeds up runtime